feat: enable building without cgo
