+++ bls.go
@@ -1,3 +1,5 @@
+//+build cgo
+
 package ffi
 
 import (
@@ -8,33 +10,6 @@ import (
 // #cgo pkg-config: ${SRCDIR}/filecoin.pc
 // #include "./filecoin.h"
 import "C"
-
-// SignatureBytes is the length of a BLS signature
-const SignatureBytes = 96
-
-// PrivateKeyBytes is the length of a BLS private key
-const PrivateKeyBytes = 32
-
-// PublicKeyBytes is the length of a BLS public key
-const PublicKeyBytes = 48
-
-// DigestBytes is the length of a BLS message hash/digest
-const DigestBytes = 96
-
-// Signature is a compressed affine
-type Signature [SignatureBytes]byte
-
-// PrivateKey is a compressed affine
-type PrivateKey [PrivateKeyBytes]byte
-
-// PublicKey is a compressed affine
-type PublicKey [PublicKeyBytes]byte
-
-// Message is a byte slice
-type Message []byte
-
-// Digest is a compressed affine
-type Digest [DigestBytes]byte
 
 // Hash computes the digest of a message
 func Hash(message Message) Digest {
