+++ proofs.go
@@ -1,3 +1,5 @@
+//+build cgo
+
 package ffi
 
 import (
@@ -15,37 +17,6 @@ import (
 // #cgo pkg-config: ${SRCDIR}/filecoin.pc
 // #include "./filecoin.h"
 import "C"
-
-// SortedPublicSectorInfo is a slice of PublicSectorInfo sorted
-// (lexicographically, ascending) by replica commitment (CommR).
-type SortedPublicSectorInfo struct {
-	f []PublicSectorInfo
-}
-
-// SortedPrivateSectorInfo is a slice of PrivateSectorInfo sorted
-// (lexicographically, ascending) by replica commitment (CommR).
-type SortedPrivateSectorInfo struct {
-	f []PrivateSectorInfo
-}
-
-// SealTicket is required for the first step of Interactive PoRep.
-type SealTicket struct {
-	BlockHeight uint64
-	TicketBytes [32]byte
-}
-
-// SealSeed is required for the second step of Interactive PoRep.
-type SealSeed struct {
-	BlockHeight uint64
-	TicketBytes [32]byte
-}
-
-type Candidate struct {
-	SectorID             uint64
-	PartialTicket        [32]byte
-	Ticket               [32]byte
-	SectorChallengeIndex uint64
-}
 
 // NewSortedPublicSectorInfo returns a SortedPublicSectorInfo
 func NewSortedPublicSectorInfo(sectorInfo ...PublicSectorInfo) SortedPublicSectorInfo {
@@ -79,11 +50,6 @@ func (s *SortedPublicSectorInfo) UnmarshalJSON(b []byte) error {
 	return json.Unmarshal(b, &s.f)
 }
 
-type PublicSectorInfo struct {
-	SectorID uint64
-	CommR    [CommitmentBytesLen]byte
-}
-
 // NewSortedPrivateSectorInfo returns a SortedPrivateSectorInfo
 func NewSortedPrivateSectorInfo(sectorInfo ...PrivateSectorInfo) SortedPrivateSectorInfo {
 	fn := func(i, j int) bool {
@@ -109,57 +75,6 @@ func (s SortedPrivateSectorInfo) MarshalJSON() ([]byte, error) {
 
 func (s *SortedPrivateSectorInfo) UnmarshalJSON(b []byte) error {
 	return json.Unmarshal(b, &s.f)
-}
-
-type PrivateSectorInfo struct {
-	SectorID         uint64
-	CommR            [CommitmentBytesLen]byte
-	CacheDirPath     string
-	SealedSectorPath string
-}
-
-// CommitmentBytesLen is the number of bytes in a CommR, CommD, CommP, and CommRStar.
-const CommitmentBytesLen = 32
-
-// SealPreCommitOutput is used to acquire a seed from the chain for the second
-// step of Interactive PoRep.
-type SealPreCommitOutput struct {
-	SectorID uint64
-	CommD    [CommitmentBytesLen]byte
-	CommR    [CommitmentBytesLen]byte
-	Pieces   []PieceMetadata
-	Ticket   SealTicket
-}
-
-// RawSealPreCommitOutput is used to acquire a seed from the chain for the
-// second step of Interactive PoRep.
-type RawSealPreCommitOutput struct {
-	CommD [CommitmentBytesLen]byte
-	CommR [CommitmentBytesLen]byte
-}
-
-// SealCommitOutput is produced by the second step of Interactive PoRep.
-type SealCommitOutput struct {
-	SectorID uint64
-	CommD    [CommitmentBytesLen]byte
-	CommR    [CommitmentBytesLen]byte
-	Proof    []byte
-	Pieces   []PieceMetadata
-	Ticket   SealTicket
-	Seed     SealSeed
-}
-
-// PieceMetadata represents a piece stored by the sector builder.
-type PieceMetadata struct {
-	Key   string
-	Size  uint64
-	CommP [CommitmentBytesLen]byte
-}
-
-// PublicPieceInfo is an on-chain tuple of CommP and aligned piece-size.
-type PublicPieceInfo struct {
-	Size  uint64
-	CommP [CommitmentBytesLen]byte
 }
 
 // VerifySeal returns true if the sealing operation from which its inputs were
